feat(minifier): remove unused class expression (#12618)
Also improved removable of unused class declaration.

closes #11454
